Refactor database connection, move address functions
